+++ app/Http/Controllers/watchlistController.php
@@ -11,38 +11,48 @@ use Illuminate\Support\Facades\Input;
 class watchlistController extends Controller
 {
     public function index(){
-        if(request()->has('active')){
-            $watchlist = Watchlist::whereHas('bid',function ($query){
-                $query->where('is_active',1);
-            })->where('user_id',Auth::user()->id)->get()->values()->all();
-        }
-        elseif (request()->has('ended')){
-            $watchlist = Watchlist::whereHas('bid',function ($query){
-                $query->where('is_active',0);
-            })->where('user_id',Auth::user()->id)->get()->values()->all();
+        if(!Auth::guest()) {
+            if (request()->has('active')) {
+                $watchlist = Watchlist::whereHas('bid', function ($query) {
+                    $query->where('is_active', 1);
+                })->where('user_id', Auth::user()->id)->get()->values()->all();
+            } elseif (request()->has('ended')) {
+                $watchlist = Watchlist::whereHas('bid', function ($query) {
+                    $query->where('is_active', 0);
+                })->where('user_id', Auth::user()->id)->get()->values()->all();
 
+            } else {
+                $watchlist = Watchlist::with('bid')->where('user_id', Auth::user()->id)->get()->values()->all();
+            }
+            return view('watchlist.index', compact('watchlist'));
+        }else{
+            return redirect('/register');
         }
-        else{
-            $watchlist = Watchlist::with('bid')->where('user_id',Auth::user()->id)->get()->values()->all();
-        }
-        return view('watchlist.index',compact('watchlist'));
 //        var_dump($watchlist);
     }
     public function delete(){
-        $list =Input::get('watchlist');
-        foreach ($list as $l){
-            $item = DB::table('watchlists')->where('id',$l);
-            $item->delete();
+        if(!Auth::guest()) {
+            $list = Input::get('watchlist');
+            foreach ($list as $l) {
+                $item = DB::table('watchlists')->where('id', $l);
+                $item->delete();
 
+            }
+            return redirect('watchlist');
+        }else{
+            return redirect('/register');
         }
-        return redirect('watchlist');
 
 
     }
     public function deleteall(){
-        $item=DB::table('watchlists')->where('user_id',Auth::user()->id);
-        $item->delete();
-        return redirect('watchlist');
+        if(!Auth::guest()) {
+            $item = DB::table('watchlists')->where('user_id', Auth::user()->id);
+            $item->delete();
+            return redirect('watchlist');
+        }else{
+            return redirect('/register');
+        }
     }
     //
 }
